feat: create dialogs element and messages arrays, imp props from App. Move files.
